设置支付密码

+++ app/src/main/java/com/jerry/jtakeaway/bean/Wallet.java
@@ -1,11 +1,10 @@
 package com.jerry.jtakeaway.bean;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 public class Wallet {
     private int id;
-    private BigDecimal balance;
+    private Double balance;
     private String paymentpassword;
     private String transactionid;
 
@@ -17,11 +16,11 @@ public class Wallet {
         this.id = id;
     }
 
-    public BigDecimal getBalance() {
+    public Double getBalance() {
         return balance;
     }
 
-    public void setBalance(BigDecimal balance) {
+    public void setBalance(Double balance) {
         this.balance = balance;
     }
